requirements.txt
requests
openpyxl
python-dateutil
tzdata
